Make the my location button work
It's still pretty terrible though: we have to wait an arbitrary amount of time after accepting the permission. Also: making the button look enabled after you click on it, but not after scrolling is pretty much impossible because this way you can't see if the scrolling is done by going to your position or by the user.